+++ src/main.ts
@@ -23,8 +23,9 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import { ConfigService } from '@nestjs/config';
-import * as cookieParser from 'cookie-parser';
 import { LogsService } from './logs/logs.service';
+import * as cookieParser from 'cookie-parser';
+import * as fs from 'fs';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
@@ -32,10 +33,25 @@ async function bootstrap() {
   const log: LogsService = app.get<LogsService>(LogsService);
   const host: string = config.get('app.host');
   const port: number = config.get('app.port');
-
   app.use(cookieParser());
 
+  log.info(`------- Starting application -------`);
+  log.info(`FWCloud Updater v${JSON.parse(fs.readFileSync('package.json').toString()).version} (PID=${process.pid})`);
+
   await app.listen(port,host);
-  this.log.info(`FWCloud Updater server listening on http://${host}:${port}`);
+  log.info(`Listening on http://${host}:${port}`);
+
+  fs.writeFileSync('.pid',`${process.pid}`);
+
+  function signalHandler (signal: 'SIGINT' | 'SIGTERM') {
+    log.info(`Received signal: ${signal}`);
+    fs.unlink('.pid',err => {
+      log.info(`------- Application stopped --------`);
+      app.close();
+      setTimeout(() => process.exit(0), 100);
+    });
+  }
+  process.on('SIGINT', signalHandler);
+  process.on('SIGTERM', signalHandler);
 }
 bootstrap();