+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Logger for store http requests into logs/http.log.
 - Middleware for log http requests.
+- Store pid in .pid file.
+- Npm script for stop process using the pid stored in .pid file.
+- SGTERM and SIGINT signal handlers.
 
 ## Changed
 - Improved compare versions error handling.